requirements.txt
pyserial>=3.5
aiomqtt>=2.0
luma.oled>=3.12
Pillow>=10.0
# Installed only on ARM targets (Raspberry Pi); skipped on x86 dev machines
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"
mh-z19>=1.0
fastapi>=0.111
uvicorn[standard]>=0.29
